pyproject.toml
[tool.poetry]
name = "test-task-webtronics-fastapi"
version = "0.1.0"
description = ""
authors = ["Leonid Sliusar <loslyusar@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task_webtronics_fastapi"}]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.17"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
bcrypt = "^4.0.1"
asyncpg = "^0.27.0"
fastapi = {extras = ["all"], version = "^0.99.0"}
pytest-asyncio = "^0.21.0"
asyncmock = "^0.4.2"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.11.1"
asynctest = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"